update dialogue system
in the beginning of the implementation, i tried to use graph view to make the system, but i found there are a lot of knowledge that i haven't learned, so it would be difficult to do it. then i decided to use scriptable object. I will learn how to use graph view when i have time.